test: 테스트 구조 변경 및 누락 테스트 추가
