--- src/test/java/roomescape/apitest/AdminReservationTimeApiTest.java
@@ -1,32 +1,0 @@
-package roomescape.apitest;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import java.util.HashMap;
-import java.util.Map;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationTimeController;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class AdminReservationTimeApiTest {
-    @Autowired
-    private ReservationTimeController reservationController;
-
-    @Test
-    @DisplayName("테마 이름이 null이면 상태코드 400을 반환한다.")
-    void 요청_이름_null_테스트() {
-        Map<String, String> params = new HashMap<>();
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/admin/times")
-                .then().log().all()
-                .statusCode(400);
-    }
-}
