+++ src/test/java/roomescape/apitest/admin/AdminReservationTimeApiTest.java
@@ -1,21 +1,20 @@
-package roomescape;
+package roomescape.apitest.admin;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.HashMap;
 import java.util.Map;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
-import static org.hamcrest.Matchers.is;
-import static roomescape.config.FixedClockConfig.FUTURE_DATE;
-
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class TimeTest {
+class AdminReservationTimeApiTest {
+
     @Test
-    void 시간_관리_API() {
+    void 시간_관리자_API() {
         Map<String, String> params = new HashMap<>();
         params.put("startAt", "19:00");
 
@@ -27,36 +26,21 @@ public class TimeTest {
                 .statusCode(201);
 
         RestAssured.given().log().all()
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(10));
-
-        RestAssured.given().log().all()
                 .when().delete("/admin/times/10")
                 .then().log().all()
                 .statusCode(204);
     }
 
     @Test
-    void 예약과_시간_연결() {
-        Map<String, Object> reservation = new HashMap<>();
-        reservation.put("name", "브라운");
-        reservation.put("date", FUTURE_DATE);
-        reservation.put("timeId", 1);
-        reservation.put("themeId", 1);
+    @DisplayName("테마 이름이 null이면 상태코드 400을 반환한다.")
+    void 요청_이름_null_테스트() {
+        Map<String, String> params = new HashMap<>();
 
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .body(reservation)
-                .when().post("/reservations")
+                .body(params)
+                .when().post("/admin/times")
                 .then().log().all()
-                .statusCode(201);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(25));
+                .statusCode(400);
     }
 }
