+++ src/test/java/roomescape/apitest/admin/AdminThemeApiTest.java
@@ -1,19 +1,15 @@
-package roomescape.apitest;
+package roomescape.apitest.admin;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationController;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 class AdminThemeApiTest {
-    @Autowired
-    private ReservationController reservationController;
 
     @Test
     void 테마_관리_API() {
