+++ src/test/java/roomescape/apitest/users/ReservationTimeApiTest.java
@@ -1,0 +1,22 @@
+package roomescape.apitest.users;
+
+import static org.hamcrest.Matchers.is;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class ReservationTimeApiTest {
+
+    @Test
+    void 시간_관리_API() {
+        RestAssured.given().log().all()
+                .when().get("/times")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(9));
+    }
+}
