+++ src/test/java/roomescape/apitest/users/ThemeApiTest.java
@@ -1,13 +1,11 @@
-package roomescape.apitest;
+package roomescape.apitest.users;
 
 import io.restassured.RestAssured;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ThemeController;
 
 import static org.hamcrest.Matchers.is;
 import static roomescape.config.FixedClockConfig.TODAY;
@@ -15,8 +13,6 @@ import static roomescape.config.FixedClockConfig.TODAY;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 class ThemeApiTest {
-    @Autowired
-    private ThemeController themeController;
 
     @Test
     void 예약_가능한_시간_조회_API() {
